Задание 2.1 перенос на ПК
